Berry Curv for Hopf with Weyl splitting
Construct the Hopf Hamiltonian with additional term for splitting of two Weyl points at critical parameters in HamiltOnLatticeWeyl.py. And calculate its eigenvalues and eigenstates.

In BerryCurvatWeyl calculate the x and z components of Berry curvature at the plane going through Weyl points (xz plane in this case). Plot the vactor field of Berry curvature in this plane.